+++ madga/__init__.py
@@ -1,5 +1,5 @@
 """MADGA — Make Django Great Again. Headless CMS for Django."""
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"
 
 default_app_config = "madga.apps.MadgaConfig"